Fix handling of links to citations
In the current version of tex4ht there’s a <span class=‘cite’> which breaks some of my clean up.

This change should fix this.

It changes slightly the expectations of the tests, but this should be fine.
It’s either whitespace, or what looks like improvements/more consistent linking.

Signed-off-by: Stefan Marr <git@stefan-marr.de>